chore(subscription): DRY subscription index in controller and test
